+++ tests/integrations/test_notebooks.py
@@ -7,6 +7,7 @@ import subprocess
 import tempfile
 import nbformat
 
+import pytest
 from unittest import TestCase
 
 default_timeout = 900
@@ -53,6 +54,7 @@ def _notebook_run(path, timeout=default_timeout):
     return nb, errors
 
 
+@pytest.mark.skipif(os.environ.get('TEST_NOTEBOOKS') != 'y', reason="Disabled notebook testing")
 class NotebooksTestCase(TestCase):
     def test_basic_example(self):
         nb, errors = _notebook_run('notebooks/examples/01 Feature Extraction and Selection.ipynb', default_timeout)